move the fetchMarketplaceProviders util inside the useFetchMarketplaceProviders file
