+++ packages/core/admin/admin/src/hooks/useFetchMarketplaceProviders/index.js
@@ -1,6 +1,21 @@
 import { useQuery } from 'react-query';
-import { useNotification } from '@strapi/helper-plugin';
-import { fetchMarketplaceProviders } from './utils/api';
+import { getFetchClient, useNotification } from '@strapi/helper-plugin';
+import qs from 'qs';
+
+const MARKETPLACE_API_URL = 'https://market-api.strapi.io';
+const { get } = getFetchClient();
+
+const fetchMarketplaceProviders = async (params = {}) => {
+  const { data } = await get(`${MARKETPLACE_API_URL}/providers`, {
+    params,
+    paramsSerializer: {
+      encode: qs.parse,
+      serialize: qs.stringify,
+    },
+  });
+
+  return data;
+};
 
 const useFetchMarketplaceProviders = (notifyLoad, params) => {
   const toggleNotification = useNotification();
